% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jac_mat.R
\name{get_jac_mat}
\alias{get_jac_mat}
\title{Create matrix for Jaccard calculation}
\usage{
get_jac_mat(list)
}
\arguments{
\item{list}{A named list of pathways and their constituent genes}
}
\value{
A matrix of overlapping genes among all pathways
}
\description{
An internal helper function to construct a matrix from the list
of pathways and their genes.
}
\references{
None.
}
\seealso{
\url{https://www.github.com/hancockinformatics/clusterPathways}
}
