% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reactome_categories_MMU_L1_L2}
\alias{reactome_categories_MMU_L1_L2}
\title{First two levels of Reactome hierarchy for all pathways in mice}
\format{
A data frame with 1699 rows and 4 columns:
\describe{
\item{id}{Reactome ID for all pathways}
\item{description}{Name of the pathway}
\item{level_1}{Highest-level name for the pathway}
\item{level_2}{Second highest-level name for the pathway}
}
}
\usage{
reactome_categories_MMU_L1_L2
}
\description{
A data frame containing all human Reactome pathways, and the corresponding
level 1 and 2 names/terms for each.
}
\keyword{datasets}
