% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reactome_genes_MMU}
\alias{reactome_genes_MMU}
\title{Genes for all mouse Reactome pathways}
\format{
A data frame with 1697 rows and 3 columns:
\describe{
\item{id}{Reactome ID for all pathways}
\item{description}{Name of the pathway}
\item{bg_genes}{All genes annotated to the pathway, as Ensembl IDs and
separated by "; "}
}
}
\usage{
reactome_genes_MMU
}
\description{
A data frame containing the constituent genes for all human Reactome pathways
}
\keyword{datasets}
