% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_reactome_pathways.R
\name{cluster_reactome_pathways}
\alias{cluster_reactome_pathways}
\title{Cluster Reactome pathways using the Jaccard index}
\usage{
cluster_reactome_pathways(input_df, reactome_levels, output_dir, width, height)
}
\arguments{
\item{input_df}{Unaltered data frame of ReactomePA results. If there is a
"direction" column with up/down for each pathway, this information will be
added to the heatmaps.}

\item{reactome_levels}{Data frame containing the first two levels for all
Reactome pathways in humans. Must have columns "id", "description",
"level_1", "level_2". If you don't have this information, the description
section contains a link to a script that will create this table for you.}

\item{output_dir}{Directory to save results. It will be created if it doesn't
already exist.}

\item{width}{Width of main output file in inches.}

\item{height}{Height of main output file in inches.}
}
\value{
A named list containing two data frames, "all_pathways" and
"rep_pathways" (see Details for more information). The function also saves
heatmaps to ".png" files in the provided output directory.
}
\description{
The following columns must be present in "input_df": "ID",
"Description", "GeneRatio", "BgRatio", "pvalue", "p.adjust", "qvalue",
"geneID", "Count".

You will likely need to play with different values of width/height to
get images that are the proper size. In some cases it may be easier to
modify the png() calls directly, and ignore the arguments here.

Link to the script to create the Reactome levels table:
\url{https://github.com/hancockinformatics/misc_R_scripts/blob/master/R_scripts/get_reactome_L1_L2.R}
}
\details{
{
The "all_pathways" table contains all pathways input to the function,
with some additional columns:

\describe{
\item{cluster}{Denotes the groups the pathways are placed within}
\item{level_1, level_2}{The highest and second-highest levels for each
pathway from the Reactoem hierarchy}
\item{candidate_genes}{The number genes annotated to the pathway that were
present in the list provided as input to ReactomePA}
\item{genes_in_pathway}{The total number of genes annotated to the pathway}
\item{gene_ratio}{\code{candidate_genes} / \code{genes_in_pathway}}
}

The "rep_pathways" table contains all of the above, plus the column
"n_pathways" which is the number of pathways from that cluster (the number in
square brackets in the "heatmap_rep_pathways_clustered.png" output image).
}
}
\references{
None.
}
\seealso{
\url{https://www.github.com/hancockinformatics/clusterReactome}
}
