% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_ratio.R
\name{get_gene_ratio}
\alias{get_gene_ratio}
\title{Title}
\usage{
get_gene_ratio(input_pathways, input_genes, species = "human")
}
\arguments{
\item{input_pathways}{Table of Reactome pathways to be clustered. Must have
the pathway ID in the first column, and description in the second.}

\item{input_genes}{Character vector of genes used to generate the
\code{input_pathways} table, e.g. a list of DE genes. Must be Ensembl IDs.}

\item{species}{Either "human" (the default) or "mouse"}
}
\value{

}
\description{
Title
}
\references{
None.
}
